+++ test-resources/db/h2/default/statements/select_kvp.sql
@@ -1,0 +1,3 @@
+select key, value
+from t_key_value_pairs
+where key = :key